requirements.txt
asgiref==3.2.5
dj-database-url==0.5.0
Django==3.0.4
djangorestframework==3.11.0
gunicorn==20.0.4
psycopg2==2.8.4
python-dotenv==0.12.0
python-memcached==1.59
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
whitenoise==5.0.1
